+++ tests/test_utils.py
@@ -178,3 +178,53 @@ def test_vector_to_shapely_latlon_polygon(tmp_path):
     with pytest.raises(ValueError, match='File contains*'):
         create_polygon(vector_file, bounds)
         utils.vector_to_shapely_latlon_polygon(vector_file)
+
+
+def test_reparse_args_burst2safe():
+    class MockArgs:
+        def __init__(self, **kwargs):
+            self.__dict__.update(kwargs)
+
+    args1 = MockArgs(granules=['granule1', 'granule2'], orbit=123)
+    with pytest.raises(ValueError, match='Cannot provide*'):
+        utils.reparse_args(args1, 'burst2safe')
+
+    args2 = MockArgs(orbit=123)
+    with pytest.raises(ValueError, match='Must provide*'):
+        utils.reparse_args(args2, 'burst2safe')
+
+    args3 = MockArgs(orbit=123, extent=['0', '0', '1', '1'], pols=['vv'], swaths=['iw1'])
+    out_args = utils.reparse_args(args3, 'burst2safe')
+    assert out_args.pols == ['VV']
+    assert out_args.swaths == ['IW1']
+    assert out_args.extent == box(*[0, 0, 1, 1])
+
+    granules = ['S1_136231_IW2_20200604T022312_VV_7C85-BURST', 'S1_136232_IW2_20200604T022315_VV_7C85-BURST']
+    args4 = MockArgs(granules=granules)
+    out_args = utils.reparse_args(args4, 'burst2safe')
+    assert out_args.granules == granules
+
+
+def test_reparse_args_burst2stack():
+    class MockArgs:
+        def __init__(self, **kwargs):
+            self.__dict__.update(kwargs)
+
+    args1 = MockArgs(rel_orb=123)
+    with pytest.raises(ValueError, match='Must provide*'):
+        utils.reparse_args(args1, 'burst2stack')
+
+    args2 = MockArgs(
+        rel_orbit=123,
+        start_date='2021-01-01',
+        end_date='2021-02-01',
+        extent=['0', '0', '1', '1'],
+        pols=['vv'],
+        swaths=['iw1'],
+    )
+    out_args = utils.reparse_args(args2, 'burst2stack')
+    assert out_args.start_date == datetime.fromisoformat('2021-01-01')
+    assert out_args.end_date == datetime.fromisoformat('2021-02-01')
+    assert out_args.pols == ['VV']
+    assert out_args.swaths == ['IW1']
+    assert out_args.extent == box(*[0, 0, 1, 1])