+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [PEP 440](https://www.python.org/dev/peps/pep-0440/)
 and uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.1]
+
+### Fixed
+* Bug in CLI parsing of granule case for `burst2safe`.
+
+### Removed
+* `--pols` as a required argument for `burst2safe` and `burst2stack`. Default value is now `VV`.
+
 ## [0.4.0]
 
 ### Added